Move glue.rs to separate folder for fmt
